warehouse_server/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pydantic==1.10.5
python-decouple==3.7
starlette==0.25.0
kafka_python
requests==2.28.0
